requirements.txt
django<4.0
djangorestframework==3.14.0
gunicorn==21.2.0
# psycopg2-binary==2.9.6
django-environ==0.10.0
pillow==10.0.0
drf-spectacular==0.26.3
drf-extra-fields==3.5.0
pytest==7.4.0
pytest-django==4.5.2
setuptools==68.0.0
djangorestframework-simplejwt==5.2.2
faker==19.2.0
django-debug-toolbar==4.2.0
django-filter==23.2
celery==5.3.1
redis==5.0.0
django-celery-results==2.5.1
django-cleanup==8.0.0